removed unused exports, adjusted data.sql and reimplemented the data model for carts, orders, ordered products + DTOs, also added a simple product for creating and updating (not displaying)

+++ backend/gameverse_backend/src/main/java/com/waff/gameverse_backend/dto/ProductDto.java
@@ -42,7 +42,7 @@ public class ProductDto {
 
     /** The tax rate applicable to the product. */
     @Positive
-    private Byte tax;
+    private Integer tax;
 
     /** The available stock quantity of the product. */
     @PositiveOrZero
